+++ CHANGELOG.md
@@ -1,5 +1,37 @@
 # CHANGELOG
 
+## 2.4.7 (2026-08-29)
+
+* Fixed [a longstanding issue][issue-406] on Windows where Alda player processes
+  were unintentionally coupled to the terminal that started them. The behavior
+  on Windows now matches the Unix/Linux behavior, where, after running an `alda
+  play` command and closing the terminal, playback continues until you run `alda
+  stop` or `alda shutdown`.
+
+  Thanks to [userjuma] for the fix!
+
+## 2.4.6 (2026-08-19)
+
+* Re-adds the `alda-code` Lisp function from Alda 1.x that takes a string and
+  parses it as Alda source:
+
+  ```alda
+  (alda-code "piano: c d e")
+
+  # This is equivalent to:
+  piano: c d e
+  ```
+
+  Thanks to [userjuma] for the contribution!
+
+## 2.4.5 (2026-08-14)
+
+* Fixed [an issue][issue-421] where, when running the Alda REPL in "server only"
+  mode (`alda repl -s`), interrupting the process with Ctrl-C was inadvertently
+  killing the background player processes.
+
+  Thanks to [Samyra312007] for the fix!
+
 ## 2.4.4 (2026-08-08)
 
 * Fixed [a timing bug in the REPL][issue-421] where a trailing rest at the end
@@ -516,6 +548,7 @@ the [Alda 2 migration guide][migration-guide]!
 [CodeForBeauty]: https://github.com/CodeForBeauty
 [SomedayToday]: https://github.com/SomedayToday
 [Samyra312007]: https://github.com/Samyra312007
+[userjuma]: https://github.com/userjuma
 
 [alda-import-blog-post]: https://blog.djy.io/musicxml-import-and-another-new-alda-features/
 [midi-channel-assignment-discussion]: https://github.com/alda-lang/alda/discussions/447
@@ -525,6 +558,8 @@ the [Alda 2 migration guide][migration-guide]!
 [midi-channel-attribute]: ./doc/attributes.md#midi-channel
 [issue-401]: https://github.com/alda-lang/alda/issues/401
 [issue-405]: https://github.com/alda-lang/alda/issues/405
+[issue-406]: https://github.com/alda-lang/alda/issues/406
 [issue-415]: https://github.com/alda-lang/alda/issues/415
 [issue-416]: https://github.com/alda-lang/alda/issues/416
+[issue-417]: https://github.com/alda-lang/alda/issues/417
 [issue-421]: https://github.com/alda-lang/alda/issues/421